app/requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.2.0
pillow==10.2.0
openai==1.12.0
python-dotenv==1.0.1
plotly==5.19.0
streamlit-extras==0.3.6
validators==0.22.0
scikit-learn==1.4.0
